common_requirements.txt
python-dotenv
openai
transformers
accelerate
einops
bitsandbytes
torch
huggingface-hub
llama-index
llama-index-llms-ollama
llama-index-embeddings-ollama
langchain
langchain-core
langchain-community
sentence-transformers
pypdf
PyPDF2
ollama
fastapi
uvicorn
streamlit
faiss-cpu
tqdm
numpy
